+++ CHANGELOG.md
@@ -11,6 +11,9 @@ CHANGELOG
 - Fetch version information from the Homebrew JSON API for CLIs installed using `brew`.
   [#3290](https://github.com/pulumi/pulumi/pull/3290)
 
+- Support renaming stack projects via `pulumi stack rename`.
+  [#3292](https://github.com/pulumi/pulumi/pull/3292)
+
 ## 1.2.0 (2019-09-26)
 
 - Support emitting high-level execution trace data to a file and add a debug-only command to view trace data.