+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+## 1.3.0 (2019-10-09)
+
 - Propagate resource inputs to resource state during preview, including first-class unknown values. This allows the
   preview to better estimate the state of a resource after an update, including property values that were populated
   using defaults calculated by the provider.